feat: add createRequests function
This function takes in a query requesting more datapoints than is supported by
the API and splits it into multiple requests of size
`API_MAX_DATAPOINTS_PER_REQUEST` or however many datapoints is left before
reaching the desired end time. Each request is Promise and an array
containing all of these is returned in the end.