feat: add function mergeResponses
Given that we are working with multiple responses but want the operation
of fetching more than 10000 datapoints to be seamless for the user of
the sdk we will merge the responses into one. The
arrays containing actual datapoints are treated specially, as they are
concatinated and sorted instead of using `mergeWith`'s default
behaviour.

Note that the sorting adds a substantial overhead when working with
large amounts of datapoints (> 250 000)