feat!: change loginWithOAuth capabilities(#496)
This PR introduces changes in authentication flow. With changes in it, loginWithOAuth method handles redirect action (handles token acquiring if it is in URI) and returns true if the token has been successfully acquired. It allows users to use authenticate method only as a separate action for token acquiring (user will be redirected to authenticate via IdP after an attempt to get token silently if it's possible). That's why few breaking changes are present:

- loginWithOAuth – becomes async, and now user can provide onHandleRedirectError to handle the redirect query error in case it is appear
- Silent login doesn’t work anymore for legacy auth flow (means via Google IdP) because of X-Frame-Options: DENY header from the Google IdP side. So there is no chance to resolve client.authenticate() without being redirecting (except the case, when the redirect was handled in loginWithOAuth method before call authenticate)
- signInType changed for the Azure AD authentication flow. Now, user able to specify login request parameters, such as prompt for instance.
- onNoProjectAvailable callback has been added for Azure AD authentication flow for handling case, when token has been successfully acquired from IdP, but user don't have access to any project in provided cluster.