fix: clone response in parallel
In node-fetch < 3.0 all large node responses will hang i cloned. And we need to clone the response to be able to fallback to text if the response is not a json (we are defauting that all responses are json).
So first solution is to run the clone in parallel (https://github.com/node-fetch/node-fetch\#custom-highwatermark)