chore(stable): fix tests failing due to spaces being deleted
There is a cleanup job that deletes old "unused" spaces, this cleanup
job works as long as the resources being used are not being reused
across test runs, as then they might be detected as "unused" in the next
run while still being used in the same run. If this happens there is a
race condition where the resource can be deleted by one test suite while
another test suite is running tests on it.

To fix this we need to ensure resources are not being reused across test
runs, such as by ensuring that a test run always uses "fresh"
resources.