feat(auth-4121): add support for sessionAcl proto type (#1296)
The access-management PR
https://github.com/cognitedata/fusion/pull/13753, tried to introduce a
change to provide predefined capability groups. Since the SDK didn't
expose the sessionsAcl type we had to rely on defining a type within the
PR which is undesirable, preferably the SDK would've supported it. This
change introduces the proto definitions found in api_acls:
https://github.com/cognitedata/infrastructure/blob/master/lib/proto/api_acls/cogmono/api_acls/sessions_acl/sessions_acl.proto.

Ticket: https://cognitedata.atlassian.net/browse/AUTH-4121

Note: this PR was made with cursor :rocket: